feat:Cleanup warning
Description:清理代码clang everything级别的编译警告。理论上不会影响到源代码逻辑.

Log:无

Issue:清理警告

+++ frame/cundocommands.cpp
@@ -1477,7 +1477,7 @@ void CSendToBackCommand::undo()
     }
     //重置保存的最大z值
     QList<QGraphicsItem *> allItems = myGraphicsScene->items();
-    int maxZvalue = 0;
+    qreal maxZvalue = 0;
     for (int i = allItems.size() - 1; i >= 0; i--) {
         QGraphicsItem *allItem = allItems.at(i);
         if (allItem->type() <= QGraphicsItem::UserType || allItem->type() >= EGraphicUserType::MgrType) {
